Separate dockyard code from normal worker buildings.
