+++ jsettlers.graphics/src/main/java/jsettlers/graphics/map/PlacementBuilding.java
@@ -14,7 +14,6 @@
  *******************************************************************************/
 package jsettlers.graphics.map;
 
-import java.util.ArrayList;
 import java.util.List;
 
 import jsettlers.common.buildings.EBuildingType;
@@ -22,7 +21,6 @@ import jsettlers.common.buildings.IBuilding;
 import jsettlers.common.buildings.IBuildingMaterial;
 import jsettlers.common.mapobject.EMapObjectType;
 import jsettlers.common.mapobject.IMapObject;
-import jsettlers.common.material.EMaterialType;
 import jsettlers.common.material.EPriority;
 import jsettlers.common.player.IPlayer;
 import jsettlers.common.player.IPlayer.DummyPlayer;
@@ -139,10 +137,5 @@ public class PlacementBuilding implements IBuilding, IBuilding.IMill {
 	@Override
 	public boolean cannotWork() {
 		return false;
-	}
-
-	@Override
-	public ArrayList<EMaterialType> getRemainingOrder() {
-		return this.getRemainingOrder();
 	}
 }
