+++ jsettlers.logic/src/main/java/jsettlers/logic/map/grid/partition/manager/manageables/interfaces/IWorkerRequestBuilding.java
@@ -14,13 +14,10 @@
  *******************************************************************************/
 package jsettlers.logic.map.grid.partition.manager.manageables.interfaces;
 
-import java.util.ArrayList;
-
 import jsettlers.common.buildings.EBuildingType;
 import jsettlers.common.buildings.IBuilding;
 import jsettlers.common.mapobject.EMapObjectType;
 import jsettlers.common.material.EMaterialType;
-import jsettlers.common.movable.EMovableType;
 import jsettlers.common.player.IPlayerable;
 import jsettlers.common.position.ILocatable;
 import jsettlers.common.position.ShortPoint2D;
@@ -76,11 +73,5 @@ public interface IWorkerRequestBuilding extends IPlayerable, ILocatable, IBuildi
 
     EMaterialType getOrderedMaterial();
 
-	void setOrder(EMaterialType[] list, EMovableType type);
-
-	ArrayList<EMaterialType> getRemainingOrder();
-
 	void reduceOrder();
-
-	void buildShipAction();
 }
